+++ src/pages/landing/components/sections/FinalCtaSection.tsx
@@ -1,0 +1,62 @@
+import { useNavigate } from "react-router-dom";
+import { ROUTES } from "@/constants/routes";
+
+const FinalCtaSection = () => {
+  const navigate = useNavigate();
+  const goSignup = () => navigate(ROUTES.SIGNUP);
+
+  return (
+    <section className="relative overflow-hidden grain py-36 mesh-bg-strong alive">
+      <div className="relative max-w-4xl mx-auto px-6 text-center reveal">
+        <div className="relative mx-auto mb-10 w-[180px] h-[180px]">
+          <div
+            className="absolute -inset-12 rounded-full opacity-70"
+            style={{
+              background:
+                "radial-gradient(closest-side, rgba(167,139,250,0.55), transparent 70%)",
+              filter: "blur(28px)",
+            }}
+          />
+          <div className="iri-sphere relative w-full h-full float-a" />
+        </div>
+
+        <h2 className="text-[44px] md:text-[64px] font-extrabold tracking-tighter2 text-ink-900 leading-[1.05]">
+          지금, 첫 결정을 <span className="why-grad">기록해보세요</span>
+        </h2>
+        <p className="mt-6 text-[18px] text-ink-500 max-w-xl mx-auto leading-[1.65]">
+          5분이면 팀 셋업이 끝납니다. 카드 등록 없이 무료로 체험해보세요.
+        </p>
+        <div className="mt-10 flex flex-wrap items-center justify-center gap-3">
+          <button
+            type="button"
+            onClick={goSignup}
+            className="btn-primary h-[56px] px-8 rounded-full text-[15px] font-semibold inline-flex items-center gap-2"
+          >
+            무료로 시작하기
+            <svg width="16" height="16" viewBox="0 0 24 24" fill="none">
+              <title>화살표</title>
+              <path
+                d="M5 12h14m-6-6 6 6-6 6"
+                stroke="currentColor"
+                strokeWidth="2.2"
+                strokeLinecap="round"
+                strokeLinejoin="round"
+              />
+            </svg>
+          </button>
+          <button
+            type="button"
+            className="btn-ghost h-[56px] px-7 rounded-full text-[15px] font-semibold"
+          >
+            팀 데모 요청하기
+          </button>
+        </div>
+        <p className="mt-7 text-[12.5px] text-ink-300">
+          신용카드 없이 시작 · 5명까지 무료 · 1분 안에 회의 분석 시작
+        </p>
+      </div>
+    </section>
+  );
+};
+
+export default FinalCtaSection;
